+++ .env.example
@@ -12,8 +12,6 @@ DB_URL=jdbc:h2:../databases/rag
 
 # If using CAII, fill these in:
 CAII_DOMAIN=
-CAII_INFERENCE_ENDPOINT_NAME=
-CAII_EMBEDDING_ENDPOINT_NAME=
 
 # set this to true if you have uv installed on your system, other wise don't include this
 USE_SYSTEM_UV=true
