+++ .project-metadata.yaml
@@ -34,14 +34,6 @@ environment_variables:
       default: ""
       description: "The domain of the CAII service. Setting this will enable CAII as the sole source for both inference and embedding models."
       required: false
-    CAII_INFERENCE_ENDPOINT_NAME:
-      default: ""
-      description: "The name of the inference endpoint for the CAII service. Required if CAII_DOMAIN is set."
-      required: false
-    CAII_EMBEDDING_ENDPOINT_NAME:
-      default: ""
-      description: "The name of the embedding endpoint for the CAII service. Required if CAII_DOMAIN is set."
-      required: false
     DB_URL:
         default: "jdbc:h2:file:~/databases/rag"
         description: "Internal DB URL. Do not change."
